requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imutils==0.5.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
protobuf==5.27.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
sympy==1.12.1
tifffile==2024.5.22
tqdm==4.66.4
urllib3==2.2.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
flatbuffers==24.3.25
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imutils==0.5.4
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
protobuf==5.27.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tifffile==2024.5.22
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
